Working "project 5" with blue led issue
- Currently SPI uses PORTD PIN1 as the sclk pin which is
  which is also the blue led pin.  Need to move.

+++ project/prop/bbb/src/spi.c
@@ -30,7 +30,7 @@ static FILE * csn_value;
 static FILE * ce_value;
 static FILE * direction;
 
-int spi_init()
+uint8_t spi_init()
 {
   // Get access to the device
   if ((fd = open(DEVICE, O_RDWR)) < 0)
